Moves HTTP handling to `http` module
